WFLY-14565 Consolidate OOBSession, OOBSessionMetaData, and OOBSessionAttributes into a single object
